+++ components/DeprecationNotice.vue
@@ -1,5 +1,5 @@
 <template>
-	<div class="vue-component message-block depcrecation-notice">
+	<div class="vue-component message-block deprecation-notice">
 		<strong>Deprecated</strong>
 		<p>Please note that this {{entity}} is now deprecated with the potential to be removed in any of the next versions. It should be transitioned out of usage as soon as possible. Refrain from using it in new implementations.</p>
 	</div>
@@ -9,7 +9,7 @@
 import Utils from '../utils';
 
 export default {
-	name: 'DepcrecationNotice',
+	name: 'DeprecationNotice',
 	props: {
 		entity: String
 	},
@@ -22,7 +22,7 @@ export default {
 <style lang="scss">
 @import './base.scss';
 
-.vue-component.depcrecation-notice {
+.vue-component.deprecation-notice {
 	background-color: rgb(255,0,0,0.1);
 	border: 1px solid red;