+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - `JsonSchema`: Improved rendering for properties.
-- `ModelBuilder`: HTML code for parameters contains invalid attributes.
+- `ModelBuilder`:
+  - HTML code for parameters contains invalid attributes.
+  - Improved UX for blocks (context-sensitive cursors etc.)
 
 ## [2.6.1] - 2020-09-07