+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.7.0] - 2021-12-15
+
 ### Added
 
 - `Collection`, `Collections`, `Item`, `Items`: Added `attribution` and `basemap` options to `mapOptions` to make basemap tile layer configurable.
@@ -478,7 +480,8 @@ First stable release.
 Please see the [Releases](https://github.com/Open-EO/openeo-vue-components/releases) for changelogs prior to v1.0.0.
 
 
-[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.6.2...HEAD
+[Unreleased]: https://github.com/Open-EO/openeo-vue-components/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.6.2...v2.7.0
 [2.6.2]: https://github.com/Open-EO/openeo-vue-components/compare/v2.6.1...v2.6.2
 [2.6.1]: https://github.com/Open-EO/openeo-vue-components/compare/v2.6.0...v2.6.1
 [2.6.0]: https://github.com/Open-EO/openeo-vue-components/compare/v2.5.1...v2.6.0